[MOVE/#104] util, custom view 패키징 수정
