+++ app/src/main/java/org/sopt/santamanitto/main/MainFragment.kt
@@ -10,7 +10,7 @@ import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
 import dagger.hilt.android.AndroidEntryPoint
 import org.sopt.santamanitto.databinding.FragmentMainBinding
-import org.sopt.santamanitto.dialog.exit.ExitDialogCreator
+import org.sopt.santamanitto.view.dialog.exit.ExitDialogCreator
 import org.sopt.santamanitto.room.create.CreateRoomActivity
 import org.sopt.santamanitto.room.manittoroom.ManittoRoomActivity
 import javax.inject.Inject
