+++ app/src/main/java/org/sopt/santamanitto/room/manittoroom/fragment/FinishFragment.kt
@@ -13,8 +13,8 @@ import org.sopt.santamanitto.R
 import org.sopt.santamanitto.databinding.FragmentManittoRoomFinishBinding
 import org.sopt.santamanitto.databinding.LayoutFinishBinding
 import org.sopt.santamanitto.databinding.LayoutResultBinding
-import org.sopt.santamanitto.dialog.RoundDialogBuilder
-import org.sopt.santamanitto.dialog.exit.ExitDialogCreator
+import org.sopt.santamanitto.view.dialog.RoundDialogBuilder
+import org.sopt.santamanitto.view.dialog.exit.ExitDialogCreator
 import org.sopt.santamanitto.room.manittoroom.ManittoRoomViewModel
 import org.sopt.santamanitto.room.manittoroom.ResultAdapter
 import org.sopt.santamanitto.view.setLayoutHeight
