--- app/src/main/java/org/sopt/santamanitto/view/BindingAdaptors.kt
@@ -1,12 +1,0 @@
-package org.sopt.santamanitto.view
-
-import android.view.View
-import androidx.databinding.BindingAdapter
-
-
-@BindingAdapter("layout_height")
-fun setLayoutHeight(view: View, height: Int) {
-    val layoutParams = view.layoutParams
-    layoutParams.height = height
-    view.layoutParams = layoutParams
-}
